Move SiblingsGroup into the content app
