+++ content/webapp/pages/page.tsx
@@ -8,7 +8,7 @@ import PageHeader from '@weco/common/views/components/PageHeader/PageHeader';
 import VideoEmbed from '@weco/common/views/components/VideoEmbed/VideoEmbed';
 import { UiImage } from '@weco/common/views/components/Images/Images';
 import { Page as PageType } from '../types/pages';
-import { SiblingsGroup } from '@weco/common/model/siblings-group';
+import { SiblingsGroup } from '../types/siblings-group';
 import {
   headerBackgroundLs,
   landingHeaderBackgroundLs,
