+++ content/webapp/services/prismic/fetch/pages.ts
@@ -21,7 +21,7 @@ import {
   guidesFields,
 } from '@weco/common/services/prismic/fetch-links';
 import { Page } from '@weco/common/model/pages';
-import { SiblingsGroup } from '@weco/common/model/siblings-group';
+import { SiblingsGroup } from '../../../types/siblings-group';
 
 export const fetchLinks = pagesFields.concat(
   articleSeriesFields,
